requirements.txt
appdirs==1.4.4
blessed==1.19.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-Cors==3.0.10
flyingcircus==0.1.4.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpickle==2.2.0
MarkupSafe==2.1.1
marshmallow==3.18.0
packaging==21.3
psycopg2-binary==2.9.5
pyparsing==3.0.9
python-decouple==3.6
python-dotenv==0.21.0
pytk==0.0.2.1
setuptools-scm==7.0.5
six==1.16.0
tk==0.1.0
tomli==2.0.1
typing_extensions==4.4.0
wcwidth==0.2.5
Werkzeug==2.2.2
zipp==3.10.0